Switched AuctionMark scale factor to be a double instead of a long. Removed hstore. prefix from cluster launching script
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/auctionmark-branch@2114 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ tests/frontend/edu/brown/benchmark/auctionmark/AuctionMarkBenchmarkProfile.java
@@ -76,7 +76,7 @@ public class AuctionMarkBenchmarkProfile implements JSONSerializable {
     /**
      * Data Scale Factor
      */
-    public long scale_factor;
+    public double scale_factor;
 
     /**
      * Map from table names to the number of tuples we inserted during loading
@@ -216,7 +216,7 @@ public class AuctionMarkBenchmarkProfile implements JSONSerializable {
      * 
      * @return
      */
-    public long getScaleFactor() {
+    public double getScaleFactor() {
         return (this.scale_factor);
     }
 
@@ -225,7 +225,7 @@ public class AuctionMarkBenchmarkProfile implements JSONSerializable {
      * 
      * @param scale_factor
      */
-    public void setScaleFactor(long scale_factor) {
+    public void setScaleFactor(double scale_factor) {
         assert (scale_factor > 0) : "Invalid scale factor " + scale_factor;
         this.scale_factor = scale_factor;
     }
